chore: Save and read files
Get data and the save it to a specified file.
This file holds the data to be used.

The data can be read and all the respective objects instantiated.
Little error handling is done for data reads and writes.

The file is created in the format "file_{username}".
This atleast makes it unique and more consistent, however,
this also risks wrong data file being read/write.
For instance a user logs in as 'jerry', and then logins
in as 'tom', this means that the file that will be created
will be different, though this means that different users
can run the program with their own data in it.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>